+++ src/components/event/Event.vue
@@ -9,6 +9,8 @@
 
     <schedule></schedule>
 
+    <preevents></preevents>
+
   </section>
 </template>
 
@@ -16,12 +18,14 @@
 import EventInfo from './EventInfo.vue'
 import Rule from './Rule.vue'
 import Schedule from './Schedule.vue'
+import PreEvents from './PreEvents.vue'
 export default {
   name: 'event',
   components:{
     eventinfo: EventInfo,
     rule: Rule,
     schedule:  Schedule,
+    preevents: PreEvents,
   },
 }
 </script>
